New django app authormgmt for managing authors and their info
Introduces a second django app for our application under test so
that we can test potential errors in migrations that have
dependencies over multiple apps (i.e. via ForeignKey relations).

Due to our DDD approach this shouldn't be the case in a real
application as we do not want to couple domains via ForeignKey
relations.