Separate Docker compose for production and develop
We are using the trick of resolivng "db.curiescience.com" to
different IPs depending if we are deploying to a local develop VM,
a stageing environment with multiple VMs or the cloud.

This is mostly done via `etc/hosts` inside the container, the
staging VM or an actual DNS server in the cloud.

It is transparent to our application.